Flatten package structure to make most classes package private.
